requirements.txt
 playwright>=1.40.0
openai>=1.10.0
python-dotenv>=1.0.0